OPCT-11: fix report-progress counters on 4.12+ (#34)
https://issues.redhat.com/browse/OPCT-11

Starting on OCP 4.12 the report-progress counters, parsed from the
output of openshift-tests binary, are broken due to the change in the
format of counters on messages for `started` lines.

- From (4.11-)
```
started: (14/3468/3476) "<test_name>"
```

- To (4.12+)
```
started: 14/3468/3476 "<test_name>"
```

As this is not the first issue when parsing the counters, and in 4.12
the total counter is also being presented, we are changing the approach
of parsing the results.

Starting now, we are changing the strategy of those counters to keep
parsing the outputs from openshift-tests binary, but not using its
counters, only the statuses lines
(`started`,`passed`,`failed`,`skipped`).

We are now appending to the `msg` field when the plugin is running, the
details of each counter in the following format: `T/C/P/F/S=N/N/N/N/N`*
to help on the debug. The letters mean:
- T: Total tests
- C: Completed tests
- P: Passed tests
- F: Failed tests
- S: Skipped tests

Sample results of the new format (running limited tests / dev mode):

```
$ KUBECONFIG=$PWD/.opct-4120/clusters/opct-4120/auth/kubeconfig oc get clusterversion
NAME      VERSION   AVAILABLE   PROGRESSING   SINCE   STATUS
version   4.12.0    True        False         3h8m    Cluster version is 4.12.0

$ KUBECONFIG=$PWD/.opct-4120/clusters/opct-4120/auth/kubeconfig $CLI_BIN_PATH run -w --dev-count=30
(...)
Thu, 26 Jan 2023 13:47:44 -03> Global Status: running
JOB_NAME                           | STATUS     | RESULTS    | PROGRESS                  | MESSAGE                                           
05-openshift-cluster-upgrade       | complete   |            | 0/0 (0 failures)          | waiting for post-processor...                     
10-openshift-kube-conformance      | complete   |            | 29/359 (0 failures)       | waiting for post-processor...                     
20-openshift-conformance-validated | running    |            | 29/3454 (0 failures)      | status=running=T/C/P/F/S=3454/29/6/0/23           
99-openshift-artifacts-collector   | running    |            | 0/0 (0 failures)          | status=waiting-for=20-openshift-conformance-validated=(0/-3441/0)=[0/1080]

$ KUBECONFIG=$PWD/.opct-4120/clusters/opct-4120/auth/kubeconfig $CLI_BIN_PATH results -p 20-openshift-conformance-validated 202301261640_sonobuoy_faa2fac5-7f0e-46d2-8334-1e8bf3eda1e0.tar.gz
Plugin: 20-openshift-conformance-validated
Status: failed
Total: 33
Passed: 9
Failed: 1
Skipped: 23

Failed tests:
[sig-arch] Monitor cluster while tests execute

```