OPCT-245: collect etcdfio data from multiple nodes (#46)
Run tests recommended step to measure disks for etcd with FIO. The step
will run into two nodes for each role (master and worker), collect the
data and the artifact collector plugin will send it to the aggregator
server.

It can help the reviewer of the Partner Validation process and prevent
back and forth in steps that can be automated.

https://issues.redhat.com/browse/OPCT-245